% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfit.R
\name{adfit}
\alias{adfit}
\title{Constructor for the \code{adfit} (Automatic differentiation AD) class}
\usage{
adfit(x)
}
\arguments{
\item{x}{Output list from \code{\link[=sample_admb]{sample_admb()}}}
}
\value{
An object of class \code{adfit}
}
\description{
Constructor for the \code{adfit} (Automatic differentiation AD) class
}
