% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adnuts.R
\docType{package}
\name{adnuts}
\alias{adnuts}
\alias{adnuts-package}
\title{adnuts: No-U-turn sampling for Template Model Builder and AD Model Builder}
\description{
The \pkg{ss3sim} \R package is designed to facilitate rapid, reproducible,
and flexible simulation with the widely-used Stock Synthesis 3 (SS3)
statistical catch-at-age stock assessment framework.
}
