% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.adfit}
\alias{as.data.frame.adfit}
\title{Convert object of class adfit to data.frame. Calls
\code{\link{extract_samples}}}
\usage{
\method{as.data.frame}{adfit}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_rwm}}}

\item{row.names}{Ignored}

\item{optional}{Ignored}

\item{...}{Ignored}
}
\value{
A data frame with parameters as columns and samples as
  rows.
}
\description{
Convert object of class adfit to data.frame. Calls
\code{\link{extract_samples}}
}
\details{
This calls the default settings of
  \code{\link{extract_samples}}, no warmup samples and no
  column for the log-posterior (lp__). Use this function
  directly for finer control.
}
