% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{as.shinystan.tmb}
\alias{as.shinystan.tmb}
\title{Convert TMB output from \link{\code{run_mcmc}} into a \code{shinystan}
object.}
\usage{
as.shinystan.tmb(tmb.fit)
}
\arguments{
\item{tmb.fit}{Output list from \link{\code{run_mcmc}} for any of the
three algorithms.}
}
\value{
An S4 object of class shinystan. Depending on the algorithm
  used, this list will have slight differences.
}
\description{
Convert TMB output from \link{\code{run_mcmc}} into a \code{shinystan}
object.
}
\details{
The shinystan packages provides several conversion functions
  for objects of different types, such as stanfit classes (Stan ouput)
  and simple arrays. For the latter, option NUTS information, such as
  \code{sampler_params} can be passed. This function essentially extends
  the functionality of \code{as.shinystan} to work specifically with TMB
  MCMC lists. The user can thus explore their TMB model with
  \code{launch_shinystan(as.shinystan.tmb(tmb.fit))} in the same way
  that Stan models are examined.
}
\seealso{
launch_shinystan_tmb
}
