% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfit.R
\name{as_data_frame_adfit}
\alias{as_data_frame_adfit}
\title{Convert object of class \code{adfit} to a \code{data.frame}}
\usage{
as_data_frame_adfit(x)
}
\arguments{
\item{x}{Output list from \code{\link[=sample_admb]{sample_admb()}}}
}
\value{
A data frame with parameters as columns and samples as
rows.
}
\description{
Convert object of class \code{adfit} to a \code{data.frame}
}
\details{
This is just a 1-1 wrapper for \code{\link[=extract_samples]{extract_samples()}}
}
