% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{benchmark_metrics}
\alias{benchmark_metrics}
\title{Calculate gradient timings on a model for different metrics}
\usage{
benchmark_metrics(obj, times = 1000, metrics = NULL, model_name = NULL)
}
\arguments{
\item{obj}{A TMB object}

\item{times}{How many evaluations to do}

\item{metrics}{A character vector of different metrics to benchmark}

\item{model_name}{An optional character name for the model, if
NULL will pull from the DLL name}
}
\value{
A data.frame containing the median gradient timing
  (time), the percent sparsity of \eqn{Q} and the dimension of
  the model (npar).
}
\description{
Calculate gradient timings on a model for different metrics
}
