% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-admb-version.R
\name{check_admb_version}
\alias{check_admb_version}
\title{Check that the  model is compiled with the right version of ADMB which is
12.0 or later by default}
\usage{
check_admb_version(
  exe = NULL,
  check_path_env = TRUE,
  min_version = 12,
  warn = TRUE
)
}
\arguments{
\item{exe}{Executable filename, compiled using ADMB}

\item{check_path_env}{Logical. If \code{TRUE}, extract the full path for the
simple executable name given by \code{exe} from the $PATH environment
variable and use that for \code{exe}. If \code{FALSE}, use \code{exe} as the full
path executable name.}

\item{min_version}{Minimum valid version (numeric). Defaults to 12.0.}

\item{warn}{Logical. If \code{TRUE}, show warnings}
}
\value{
The version number as a character string, invisibly
}
\description{
Check that the  model is compiled with the right version of ADMB which is
12.0 or later by default
}
\details{
Some functionality of packages \link{adnuts} is embedded in the ADMB
source code so that when a model is compiled it is contained in the model
executable. If this code does not exist, \code{adnuts} will fail. The solution
is to update ADMB and recompile the model.
}
