% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{combine_fits}
\alias{combine_fits}
\title{Combine multiple fits as returned from \code{sample_tmb} or
\code{sample_admb} run as a single chain.}
\usage{
combine_fits(fits)
}
\arguments{
\item{fits}{A list of fits, each having a single chain}
}
\value{
A merged fit across chains.
}
\description{
Combine multiple fits as returned from \code{sample_tmb} or
\code{sample_admb} run as a single chain.
}

