% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.as.shinyadnuts}
\alias{.as.shinyadnuts}
\title{Convert adnuts fit (named list) into a \code{shinystan} object.}
\usage{
.as.shinyadnuts(fit)
}
\arguments{
\item{fit}{Output list from \code{\link[=sample_admb]{sample_admb()}}.}
}
\value{
An S4 object of class shinystan. Depending on the algorithm
used, this list will have slight differences.
}
\description{
Convert adnuts fit (named list) into a \code{shinystan} object.
}
\details{
The shinystan packages provides several conversion functions
for objects of different types, such as stanfit classes (Stan ouput)
and simple arrays. For the latter, option NUTS information, such as
\code{\link[=sampler_params]{sampler_params()}} can be passed. This function essentially extends
the functionality of \code{\link[=as.shinystan]{as.shinystan()}} to work specifically with
fits from adnuts (TMB or ADMB). The user can thus explore their model
with \link{launch_shinystan(.as.shinyadnuts(fit))} in the same way
that Stan models are examined.
}
\seealso{
\code{\link[=launch_shinyadmb]{launch_shinyadmb()}}
}
