% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_console_printing}
\alias{.check_console_printing}
\title{Check if the session is interactive or Rstudio which has
implications for parallel output}
\usage{
.check_console_printing(parallel)
}
\arguments{
\item{parallel}{Boolean whether chain is executed in parallel
mode or not.}
}
\value{
Boolean whether output should be printed to console
  progressively, or saved to file and printed at the end.
}
\description{
Check if the session is interactive or Rstudio which has
implications for parallel output
}
\details{
When using RStudio and RGui, the parallel output does
  not show on the console. As a workaround it is captured in
  each cluster into a file and then read in and printed.
}
