% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_model_path}
\alias{.check_model_path}
\title{Check that the file can be found}
\usage{
.check_model_path(model, path)
}
\arguments{
\item{model}{Model name without file extension}

\item{path}{Path to model folder, defaults to working}
}
\description{
Check that the file can be found
}
