% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_psv}
\alias{.get_psv}
\title{Read in PSV file}
\usage{
.get_psv(model, path)
}
\arguments{
\item{model}{The name of the executable (may have a prepended ./)}

\item{path}{The path in which the PSV file is located}
}
\value{
The contents of the PSV file as read in by [R2admb::read_psv()]
}
\description{
Read in PSV file
}
