% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{.get_Q}
\alias{.get_Q}
\title{Get the joint precision matrix Q from an optimized TMB or RTMB obj.}
\usage{
.get_Q(obj)
}
\arguments{
\item{obj}{An optimized TMB or RTMB object}
}
\value{
A sparse matrix Q
}
\description{
Get the joint precision matrix Q from an optimized TMB or RTMB obj.
}
