% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{.get_Qinv}
\alias{.get_Qinv}
\title{Get the joint covariance Sigma from an optimized TMB or RTMB
obj without random effects.}
\usage{
.get_Qinv(obj)
}
\arguments{
\item{obj}{An optimized TMB or RTMB object}
}
\value{
A dense matrix Sigma
}
\description{
Get the joint covariance Sigma from an optimized TMB or RTMB
obj without random effects.
}
