% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.getADMBHessian}
\alias{.getADMBHessian}
\title{Read in admodel.hes file}
\usage{
.getADMBHessian(path)
}
\arguments{
\item{path}{Path to folder containing the admodel.hes file}
}
\value{
The Hessian matrix
}
\description{
Read in admodel.hes file
}
