% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_unique_names}
\alias{.make_unique_names}
\title{Function to take a character vector of parameter names and
force them to be unique by appending numbers in square
brackets as needed}
\usage{
.make_unique_names(x)
}
\arguments{
\item{x}{Character vector}
}
\description{
Function to take a character vector of parameter names and
force them to be unique by appending numbers in square
brackets as needed
}
