% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.print.mcmc.progress}
\alias{.print.mcmc.progress}
\title{Print MCMC progress to console.}
\usage{
.print.mcmc.progress(iteration, iter, warmup, chain)
}
\arguments{
\item{iteration}{The iteration of the MCMC chain.}

\item{iter}{The total iterations.}

\item{warmup}{The number of warmup iterations.}

\item{chain}{The chain being run (bookkeeping only).}
}
\value{
Nothing. Prints to message to console.
}
\description{
Print MCMC progress to console.
}
\details{
This function was modeled after the functionality provided by
the R package rstan.
}
