% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rotate_space}
\alias{.rotate_space}
\title{Update algorithm for mass matrix.}
\usage{
.rotate_space(fn, gr, M, y.cur)
}
\arguments{
\item{fn}{The current fn function.}

\item{gr}{The current gr function}

\item{M}{The new mass matrix}

\item{y.cur}{The current parameter vector in unrotated (Y) space.}
}
\description{
Update algorithm for mass matrix.
}
