% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.update_model}
\alias{.update_model}
\title{Convert model name depending on system}
\usage{
.update_model(model)
}
\arguments{
\item{model}{Model name without file extension}
}
\value{
Updated model name to use with system call
}
\description{
Convert model name depending on system
}
