% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract posterior samples from a TMB MCMC fit list.}
\usage{
extract_samples(fit.tmb, inc_warmup = FALSE)
}
\arguments{
\item{fit.tmb}{A list returned by \code{\link{run_mcmc}}.}

\item{inc_warmup}{Whether to extract the warmup samples or not
(default). Warmup samples should never be used for inference, but may
be useful for diagnostics.}
}
\value{
An invisible data.frame containing samples (rows) of each
  parameter (columns). If multiple chains exist they will be rbinded
  together.
}
\description{
Extract posterior samples from a TMB MCMC fit list.
}
