% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get.admb.cov}
\alias{get.admb.cov}
\title{Read in the ADMB covariance file.}
\usage{
get.admb.cov(model.path = getwd())
}
\arguments{
\item{model.path}{Path to model (defaults to working directory)}
}
\description{
Read in the ADMB covariance file.
}

