% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-model-executable.R
\name{get_model_executable}
\alias{get_model_executable}
\title{Find the model executable name on your filesystem}
\usage{
get_model_executable(model, loc_dir = getwd(), path_only = FALSE)
}
\arguments{
\item{model}{The name of the executable (without any extension)}

\item{loc_dir}{The path to check for existence of the executable before
checking the path}

\item{path_only}{If \code{TRUE}, ignore \code{loc_dir} and search the PATH only}
}
\value{
The executable as it should be called in the OS. It is either
just as it was given if on the \verb{$PATH} or in the \code{loc_dir} directory on a
Windows machine, or pre-pended with \verb{./} if in the \code{loc_dir} directory on a
non-Windows machine. If not found anywhere, an error will be thrown and
code execution will stop
}
\description{
Find the model executable name in the local directory (\code{loc_dir}).
If not there, search the \verb{$PATH}.
If found more than once in the \verb{$PATH}, the first one will be used and a
warning issued.
If not on the \verb{$PATH}, throw an error and stop code execution.
If using Linux or OSX and the file permissions do not allow the user to
execute, throw an error and stop code execution.
}
