% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{get_post}
\alias{get_post}
\title{Extract posterior samples from a tmbfit object}
\usage{
get_post(x, invf, parnames, array = FALSE)
}
\arguments{
\item{x}{A fitted tmbfit object}

\item{invf}{The inverse function to decorrelate the parameters}

\item{parnames}{A vector of parameter names, excluding lp__}

\item{array}{Whether to return a data.frame (default) or array
which is used in constructing other objects downstream}
}
\description{
Extract posterior samples from a tmbfit object
}
