% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-runtimes.R
\name{get_runtimes}
\alias{get_runtimes}
\title{Extract runtime, final acceptance ratio, and final stepsize
information from logffiles after running adnuts}
\usage{
get_runtimes(path)
}
\arguments{
\item{path}{}
}
\value{
A list of numeric vectors including warmup and sampling runtimes,
final acceptance ratio, and final stepsize
}
\description{
Extract runtime, final acceptance ratio, and final stepsize
information from logffiles after running adnuts
}
\details{
The information is extracted from the logfiles, 1 for each
chain
}
