% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfit.R
\name{is_adfit}
\alias{is_adfit}
\title{Does object \code{x} inherit class \code{adfit}?}
\usage{
is_adfit(x)
}
\arguments{
\item{x}{Output list from \code{\link[=sample_admb]{sample_admb()}}}
}
\value{
Logical
}
\description{
Does object \code{x} inherit class \code{adfit}?
}
