% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{launch_shinystan_tmb}
\alias{launch_shinystan_tmb}
\title{A high level wrapper to launch shinystan for a TMB MCMC list object.}
\usage{
launch_shinystan_tmb(tmb.fit)
}
\description{
A high level wrapper to launch shinystan for a TMB MCMC list object.
}
\details{
This function simply calls
  \code{launch_shinystan(as.shinystan.tmb(tmb.fit))}.
}

