% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{launch_shinytmb}
\alias{launch_shinytmb}
\title{Launch shinystan for a TMB fit.}
\usage{
launch_shinytmb(fit)
}
\arguments{
\item{fit}{A named list returned by \code{\link[=sample_tmb]{sample_tmb()}}}
}
\description{
Launch shinystan for a TMB fit.
}
\seealso{
\code{\link[=launch_shinyadmb]{launch_shinyadmb()}}
}
