% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.adfit}
\alias{plot.adfit}
\title{Plot object of class adfit}
\usage{
\method{plot}{adfit}(x, y, ...)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_admb}}}

\item{y}{Ignored}

\item{...}{Ignored}
}
\value{
Plot created
}
\description{
Plot object of class adfit
}
