% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfit.R
\name{plot_adfit}
\alias{plot_adfit}
\title{Plot object of class \code{adfit}}
\usage{
plot_adfit(x)
}
\arguments{
\item{x}{Output list from \code{\link[=sample_admb]{sample_admb()}}}
}
\value{
Nothing
}
\description{
Plot object of class \code{adfit}
}
\details{
This is just a 1-1 wrapper for \code{\link[=plot_marginals]{plot_marginals()}}
}
