% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{plot_Q}
\alias{plot_Q}
\title{Make an image plot showing the correlation (lower triangle)
and sparsity (upper triangle).}
\usage{
plot_Q(fit, Q = NULL)
}
\arguments{
\item{fit}{A fitted object}

\item{Q}{A sparse matrix. If NULL it will be extracted from
\code{fit}.}
}
\value{
A plot created by \code{Matrix::image}.
}
\description{
Make an image plot showing the correlation (lower triangle)
and sparsity (upper triangle).
}
\details{
This function is used to visualize the sparsity and
  correlation patterns of the joint model. The upper triangle
  shows whether an element is 0 (white) or not (gray), while
  the lower triangle shows the correlation calculated from
  \code{cov2cor(solve(Q))}.
}
