% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sampler-params.R
\name{plot_sampler_params}
\alias{plot_sampler_params}
\title{Plot adaptation metrics for a fitted model.}
\usage{
plot_sampler_params(
  fit,
  label_size = 12,
  label_face = c("plain", "bold", "italic"),
  ...
)
}
\arguments{
\item{fit}{A fitted object returned by \code{\link[=sample_admb]{sample_admb()}}.}

\item{...}{Arguments passed to \code{\link[=extract_sampler_params]{extract_sampler_params()}}}
}
\value{
Prints and invisibly returns a ggplot object
}
\description{
Plot adaptation metrics for a fitted model.
}
\details{
This utility function quickly plots the adaptation output of NUTS
chains.
}
\examples{
fit <- readRDS(system.file('examples', 'fit.RDS', package='adnuts'))
plot_sampler_params(fit)
}
