% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.adfit}
\alias{print.adfit}
\title{Print summary of adfit object}
\usage{
\method{print}{adfit}(fit)
}
\arguments{
\item{fit}{Fitted object from \code{\link{sample_admb}}}
}
\description{
Print summary of adfit object
}
