% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfit.R
\name{print_adfit}
\alias{print_adfit}
\title{Print summary of an \code{adfit} object}
\usage{
print_adfit(x, ...)
}
\arguments{
\item{x}{Output list from \code{\link[=sample_admb]{sample_admb()}}}

\item{...}{Absorb arguments intended for other functions}
}
\value{
Nothing
}
\description{
Print summary of an \code{adfit} object
}
