% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-admb-cov.R
\name{read_admb_cov}
\alias{read_admb_cov}
\title{Read in the ADMB covariance file.}
\usage{
read_admb_cov(path = NULL)
}
\arguments{
\item{path}{Path to model}
}
\value{
A list of named values
}
\description{
Read in the ADMB covariance file.
}
