% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-admb-hessian.R
\name{read_admb_hessian}
\alias{read_admb_hessian}
\title{Read in admodel.hes file}
\usage{
read_admb_hessian(path)
}
\arguments{
\item{path}{Path to directory containing the admodel.HES file}
}
\value{
The Hessian matrix
}
\description{
Read in admodel.hes file
}
\details{
This function reads in all of the information contained in the
admodel.hes file. Some of this is needed for relaxing the
covariance matrix, and others just need to be recorded and
rewritten to file so ADMB "sees" what it's expecting.
}
