% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-mle-fit.R
\name{read_mle_fit}
\alias{read_mle_fit}
\title{Read maximum likelihood fit for ADMB model}
\usage{
read_mle_fit(path = NULL)
}
\arguments{
\item{path}{The path containing the MLE output you wish to load (contains
a \code{par} and \code{cor} file)}
}
\value{
A list containing, MLE estimates, standard errors, covariance
and correlation matrices, and other output from ADMB
}
\description{
Read maximum likelihood fit for ADMB model
}
\details{
This is based loosely on \code{\link[r4ss:read.admbFit]{r4ss::read.admbFit()}}
Sequentially read .par file which contains model size, minimum NLL, and
maxgrad at the top.
}
