% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-psv.R
\name{read_psv}
\alias{read_psv}
\title{Read in an ADMB PSV file}
\usage{
read_psv(path, fn_psv = NULL, nms = NULL)
}
\arguments{
\item{path}{The path to the location of the PSV file}

\item{fn_psv}{The name of the psv file, with or without the extension
psv or .PSV}

\item{nms}{Column names for the output. If \code{NULL}, they will be sequential
starting with the letter 'V'}
}
\value{
A data frame with the PSV file output
}
\description{
Read in an ADMB PSV file
}
