% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate-space.R
\name{rotate_space}
\alias{rotate_space}
\title{Update algorithm for mass matrix.}
\usage{
rotate_space(fn, gr, M, y_cur)
}
\arguments{
\item{fn}{The current fn function.}

\item{gr}{The current gr function}

\item{M}{The new mass matrix}

\item{y_cur}{The current parameter vector in unrotated (Y) space.}
}
\description{
Update algorithm for mass matrix.
}
\details{
Rotation done using choleski decomposition
First case is a dense mass matrix
}
