% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-admb-sampler.R
\name{run_admb_sampler}
\alias{run_admb_sampler}
\title{Setup directories and run a sampler, either \code{\link[=sample_admb_rwm]{sample_admb_rwm()}}
or \code{\link[=sample_admb_nuts]{sample_admb_nuts()}}}
\usage{
run_admb_sampler(chain_num, path, algorithm = c("nuts", "rwm"), ...)
}
\arguments{
\item{chain_num}{The chain number}

\item{path}{The path name to append the chain number to}

\item{algorithm}{The algorithm to use, one of \code{nuts} or \code{rwm}}

\item{...}{Arguments passed to the \code{\link[=fit]{fit()}} function}
}
\value{
Output from the \code{\link[=fit]{fit()}} function
}
\description{
Setup directories and run a sampler, either \code{\link[=sample_admb_rwm]{sample_admb_rwm()}}
or \code{\link[=sample_admb_nuts]{sample_admb_nuts()}}
}
