% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{run_mcmc}
\alias{run_mcmc}
\title{MCMC sampling of TMB models}
\usage{
run_mcmc(obj, iter, algorithm = "NUTS", chains = 1, init = NULL,
  covar = NULL, lower = NULL, upper = NULL, thin = 1, ...)
}
\arguments{
\item{obj}{A TMB model object.}

\item{iter}{The number of (dependent) samples to draw.}

\item{algorithm}{A string specifiying an algorithm. Currently supported
are: \itemize{ \item{"RWM"}{the random walk Metropolis sampler}
\item{"HMC"}{the Hamiltonian sampler (see Neal 2011)}
\item{"NUTS"}{the No-U-Turn sampler (see Hoffman and Gelman 2014)} }
These algorithms require different arguments; see their help files for
more information.}

\item{init}{A list of lists containing the initial parameter vectors,
one for each chain. It is strongly recommended to initialize multiple
chains from dispersed points. The default of NULL signifies to use the
starting values present in the model (i.e., \code{obj$par}), but this
will apply to all chains.}

\item{covar}{An optional covariance matrix which can be used to improve
the efficiency of sampling. The lower Cholesky decomposition of this
matrix is used to transform the parameter space. If the posterior is
approximately multivariate normal and \code{covar} approximates the
covariance, then the transformed parameter space will be close to
multivariate standard normal. In this case the algorithm will be more
efficient, but there will be overhead in the matrix calculations which
need to be done at each step. The default of NULL specifies to not do
this transformation.}

\item{thin}{The thinning rate to apply to samples. The default of 1
means to not thin (keep all samples). Thinning usually is not
necessary for NUTS but in some cases may be useful.}

\item{...}{Further arguments to be passed to the algorithm. See help
files for the samplers for further arguments.}
}
\value{
A list containing the samples and properties of the sampler
  useful for diagnosing behavior and efficiency.
}
\description{
[BETA VERSION] Draw samples from the posterior of a TMB model using a
specified MCMC algorithm.
}
\details{
This function is a top-level wrapper designed specifically to
  work with TMB models. There are several MCMC algorithms available for
  use. The user is responsible for specifying the model properly
  (priors, starting values, desired parameters fixed, etc.), as well as
  assessing the convergence and validity of the resulting samples (e.g.,
  through the \code{coda} package) before making inference.
}
\author{
Cole Monnahan
}
\seealso{
\code{\link{run_mcmc.hmc}}, \code{\link{run_mcmc.nuts}},
  \code{\link{run_mcmc.rwm}}
}

