% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss_mcmc.R
\name{run_ss_mcmc}
\alias{run_ss_mcmc}
\title{Run an MCMC using an ADMB model, return (1) the posterior draws, MLE
fits and covariance/correlation matrices, and some MCMC convergence
diagnostics using CODA.}
\usage{
run_ss_mcmc(model.path, model.name, Nout, mcsave, burn.in, cov.user = NULL,
  init.pin = NULL, se.scale = NULL, mcscale = TRUE, mcseed = NULL,
  mcrb = NULL, mcdiag = FALSE, mcprobe = NULL, verbose = TRUE,
  extra.args = NULL, hybrid = FALSE, hyeps = NULL, hynstep = NULL,
  mceval = FALSE, estimate = FALSE)
}
\arguments{
\item{model.path}{(Character) A path to the folder containing the model. NULL
indicates the current folder.}

\item{Nout}{(Integer) The number of draws after thinning and burn in.}

\item{mcsave}{(Integer) Controls thinning of samples. Save every mcsave
value, such that 1 corresponds to keeping all draws, and 100 saving
every 100th draw.}

\item{burn.in}{(Integer) How many samples to discard from the beginning
of the chain, *after* thining. The burn in period (i.e., the first
burn.in*mcsave draws) should be at least large enough to cover dynamic
scaling.}

\item{cov.user}{(Numeric matrix) A manually defined covariance matrix (in bounded space)
to use in the Metropolis-Hastings algorithm.}

\item{init.pin}{(Numeric vector) A vector of initial values, which are written to file
and used in the model via the -mcpin option.}

\item{se.scale}{(Numeric) A value which scales all of the variances from
the MLE fit. A value of 1 indicates to use the estimated variances.}

\item{mcscale}{(Logical) Whether to use the mcscale option, which
dynamically scales the covariance matrix for efficient acceptance
ratios.}

\item{mcseed}{(Integer) Which seed (integer value) to pass ADMB. Used
for reproducibility.}

\item{mcrb}{(Integer) Which value to use in the rescale bounded
algorithm. Must be an integer from 1-9. The default NULL value disables
this feature. See the vignette for more information on this algorithm
and how to best use it.}

\item{mcdiag}{(Logical) Whether to use the \code{mcdiag} feature. This
uses an identity matrix for the covariance matrix.  #' @param mcprobe
Which value to use in the probing algorithm. The default NULL value
disables this feature. See the vignette for more information on this
algorithm and how to best use it.}

\item{verbose}{(Logical) Whether to print ADMB warnings and other
information. Useful for testing and troubleshooting.}

\item{extra.args}{(Character) A string which is passed to ADMB at
runtime. Useful for passing additional arguments to the model
executable.}

\item{hyeps}{(Numeric) The size of the leapfrog jump in the hybrid
method, with smaller values leading to smaller but more accurate
jumps. Must be a positive value.}

\item{hynstep}{(Integer) The approximate number of steps used in the
leapfrog step of the hybrid algorithm. Steps are randomly generated for
each MCMC iteration, centered around \code{hynstep}.}

\item{mode.name}{(Character) The name of the model executable. A character string,
without '.exe'.}

\item{hyrbid}{(Logical) Whether to use the Hamiltonial (hybrid)
algorithm. Default is FALSE.}
}
\value{
Returns a list containing (1) the posterior draws, (2) and
object of class 'admb', read in using the results read in using
\code{read_admb}, and (3) some MCMC convergence diagnostics using CODA.
}
\description{
Run an MCMC using an ADMB model, return (1) the posterior draws, MLE
fits and covariance/correlation matrices, and some MCMC convergence
diagnostics using CODA.
}

