% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss_mcmc.R
\name{run_ss_mcmc_parallel}
\alias{run_ss_mcmc_parallel}
\title{A wrapper for running SS models in parallel}
\usage{
run_ss_mcmc_parallel(parallel_number, Nout, mcsave, burn.in, cov.user = NULL,
  init.pin = NULL, se.scale = 1, mcscale = TRUE, mcseed = NULL,
  mcrb = NULL, mcdiag = FALSE, mcprobe = NULL, verbose = TRUE,
  extra.args = NULL, hybrid = FALSE, hyeps = NULL, hynstep = NULL,
  mceval = FALSE, estimate = FALSE)
}
\description{
A wrapper for running SS models in parallel
}

