% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_admb.R
\name{sample_admb_nuts}
\alias{sample_admb_nuts}
\title{Run a single NUTS chain for an ADMB model}
\usage{
sample_admb_nuts(path, model, iter = 2000, init = NULL, chain = 1,
  thin = 1, warmup = NULL, seed = NULL, duration = NULL,
  control = NULL, verbose = TRUE, extra.args = NULL)
}
\description{
A low level function to run a single chain. Unlikely to be used by a
user, instead prefer \code{\link{sample_admb}}
}
\seealso{
sample_admb
}
