% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{sample_admb_parallel}
\alias{sample_admb_parallel}
\title{A wrapper for running ADMB models in parallel}
\usage{
sample_admb_parallel(parallel_number, path, algorithm, ...)
}
\arguments{
\item{parallel_number}{The chain number}

\item{path}{The path name to append the chain number to}

\item{algorithm}{NUTS or RWM}

\item{...}{Arguments passed to the \code{\link[=fit]{fit()}} function}
}
\value{
Output from the \code{\link[=fit]{fit()}} function
}
\description{
A wrapper for running ADMB models in parallel
}
