% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_admb.R
\name{sample_admb_rwm}
\alias{sample_admb_rwm}
\title{Run a single random walk Metropolis chain for an ADMB model}
\usage{
sample_admb_rwm(path, model, iter = 2000, thin = 1,
  warmup = ceiling(iter/2), init = NULL, chain = 1, seed = NULL,
  control = NULL, par.names = NULL, verbose = TRUE, extra.args = NULL,
  duration = NULL, mceval = TRUE)
}
\description{
A low level function to run a single chain. Unlikely to be used by a
user, instead prefer \code{\link{sample_admb}}
}
\seealso{
sample_admb
}
