% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_inits}
\alias{sample_inits}
\title{Function to generate random initial values from a previous fit using
adnuts}
\usage{
sample_inits(fit, chains)
}
\arguments{
\item{fit}{An outputted list from \code{\link{sample_admb}}}

\item{chains}{The number of chains for the subsequent run, which
determines the number to return.}
}
\value{
A list of lists which can be passed back into
  \code{\link{sample_admb}}.
}
\description{
Function to generate random initial values from a previous fit using
adnuts
}
