% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptation.R
\name{slow_phase}
\alias{slow_phase}
\title{Check whether adaptation is in the slow phase}
\usage{
slow_phase(i, warmup, w1, w3)
}
\arguments{
\item{i}{MCMC iteration number}

\item{warmup}{Number of warmup iterations}

\item{w1}{The first adapation window (usually 75)}

\item{w3}{The last adaptation window (usually 50)}
}
\value{
Bool whether in slow phase
}
\description{
Check whether adaptation is in the slow phase
}
\details{
During the slow phase the mass matrix is updated in a series of
  expanding windows. See Stan manual on adaptation.
}

