% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.adfit}
\alias{summary.adfit}
\title{Print summary of object of class adfit}
\usage{
\method{summary}{adfit}(fit)
}
\arguments{
\item{fit}{Fitted object from \code{\link{sample_admb}}}
}
\description{
Print summary of object of class adfit
}
