% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfit.R
\name{summary_adfit}
\alias{summary_adfit}
\title{Print summary of object of class \code{adfit}}
\usage{
summary_adfit(x)
}
\arguments{
\item{x}{Output list from \code{\link[=sample_admb]{sample_admb()}}}
}
\value{
Nothing
}
\description{
Print summary of object of class \code{adfit}
}
\details{
This is just a 1-1 wrapper for \code{\link[=print_adfit]{print_adfit()}}
}
