% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-logfiles.R
\name{trim_logfile}
\alias{trim_logfile}
\title{Trim a logfile down in size by removing all the MCMC step outputs}
\usage{
trim_logfile(fn)
}
\arguments{
\item{fn}{The filename to modify}
}
\value{
Nothing
}
\description{
Trim a logfile down in size by removing all the MCMC step outputs
}
