% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-logfiles.R
\name{trim_logfiles}
\alias{trim_logfiles}
\title{Trim logfiles down in size by removing all the MCMC step outputs}
\usage{
trim_logfiles(path)
}
\arguments{
\item{path}{The path where the logfiles reside}
}
\value{
Nothing
}
\description{
Trim logfiles down in size by removing all the MCMC step outputs
}
\details{
Searches a \code{path} for all files matching \code{chain_XX_model_output.log}
where XX are digits representing chain numbers
}
