% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{update_control}
\alias{update_control}
\title{Update the control list.}
\usage{
update_control(control)
}
\arguments{
\item{control}{A list passed from \code{sample_tmb}.}
}
\value{
A list with default control elements updated by those supplied
  in \code{control}
}
\description{
Update the control list.
}

