% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{write.admb.cov}
\alias{write.admb.cov}
\title{Write a covariance matrix to admodel.cov.}
\usage{
write.admb.cov(cov.unbounded, model.path = getwd(), hbf = NULL)
}
\arguments{
\item{cov.unbounded}{The cov matrix in unbounded space.}

\item{model.path}{Path to model.}

\item{hbf}{The hybrid_bounded_flag value. Use hbf=1 for HMC.}
}
\description{
Write a covariance matrix to admodel.cov.
}

