% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_admb_cov}
\alias{write_admb_cov}
\title{Write a covariance matrix to admodel.cov.}
\usage{
write_admb_cov(cov_unbounded, path = NULL, hbf = NULL)
}
\arguments{
\item{cov_unbounded}{The cov matrix in unbounded space}

\item{hbf}{The hybrid_bounded_flag value. Use \code{hbf = 1} for HMC.}

\item{model_path}{Path to model.}
}
\description{
Write a covariance matrix to admodel.cov.
}
