% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_psv}
\alias{write_psv}
\title{Write matrix of samples to a binary .psv file.}
\usage{
write_psv(fn, samples, path = NULL)
}
\arguments{
\item{fn}{Model executable name}

\item{samples}{A matrix or data.frame of samples, each column is a
parameter, each row a sample.}
}
\description{
Write matrix of samples to a binary .psv file.
}
\details{
Useful to combine multiple MCMC runs together into a single
.psv file which can then be executed with \code{-mceval}.
}
