% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adfit}
\alias{adfit}
\title{Constructor for the "adfit" (A-D fit) class}
\usage{
adfit(fit)
}
\arguments{
\item{fit}{Fitted object from \code{\link{sample_admb}}}
}
\value{
An object of class "adfit"
}
\description{
Constructor for the "adfit" (A-D fit) class
}
