% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{as.tmbfit}
\alias{as.tmbfit}
\title{Construtor for tmbfit objects}
\usage{
as.tmbfit(x, mle, invf, metric, model = "anonymous")
}
\arguments{
\item{x}{A fitted MCMC object}

\item{mle}{A list of MLE parameters}

\item{invf}{The inverse function for the parameters}

\item{metric}{The metric used}

\item{model}{A character giving the model name}
}
\description{
Construtor for tmbfit objects
}
