% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_identifiable}
\alias{check_identifiable}
\title{Check identifiability from model Hessian}
\usage{
check_identifiable(model, path = getwd())
}
\arguments{
\item{model}{Model name without file extension}

\item{path}{Path to model folder, defaults to working directory}
}
\value{
Prints output of bad parameters and invisibly returns it.
}
\description{
Check identifiability from model Hessian
}
\details{
Read in the admodel.hes file and check the eigenvalues to
  determine which parameters are not identifiable and thus cause the
  Hessian to be non-invertible. Use this to identify which parameters
  are problematic. This function was converted from a version in the
  \code{FishStatsUtils} package.
}
