% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getADMBHessian}
\alias{.getADMBHessian}
\title{Read in admodel.hes file}
\usage{
.getADMBHessian(path, full = FALSE)
}
\arguments{
\item{path}{Path to folder containing the admodel.hes file}

\item{full}{Whether to return just the Hessian (FALSE -
default) or the full}
}
\value{
The Hessian matrix
}
\description{
Read in admodel.hes file
}
