% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.print.mcmc.timing}
\alias{.print.mcmc.timing}
\title{Print MCMC timing to console}
\usage{
.print.mcmc.timing(runtime)
}
\arguments{
\item{runtime}{Time of total in seconds.}
}
\value{
Nothing. Prints message to console.
}
\description{
Print MCMC timing to console
}
\details{
This function was modeled after the functionality provided by
the R package \link{rstan}
}
