% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_sampler_params}
\alias{extract_sampler_params}
\title{Extract sampler parameters from a fit}
\usage{
extract_sampler_params(fit, inc_warmup = FALSE)
}
\arguments{
\item{fit}{A list returned by \code{sample_admb} or \code{sample_tmb}.}

\item{inc_warmup}{Whether to extract the warmup samples or not
(default). Warmup samples should never be used for inference, but may
be useful for diagnostics.}
}
\value{
An invisible data.frame containing samples (rows) of each
  parameter (columns). If multiple chains exist they will be rbinded
  together.
}
\description{
Extract sampler parameters from a fit
}
