% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{launch_shinyadmb}
\alias{launch_shinyadmb}
\title{Launch shinystan for an ADMB fit.}
\usage{
launch_shinyadmb(fit)
}
\arguments{
\item{fit}{A named list returned by \code{sample_admb}.}
}
\description{
Launch shinystan for an ADMB fit.
}
\seealso{
\code{launch_shinytmb}
}
