% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{launch_shinyadmb}
\alias{launch_shinyadmb}
\title{A high level wrapper to launch shinystan for a ADMB fit.}
\usage{
launch_shinyadmb(fit)
}
\description{
A high level wrapper to launch shinystan for a ADMB fit.
}
\details{
This function simply calls
  \code{launch_shinystan(as.shinystan.tmb(tmb.fit))}.
}
