% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{launch_shinytmb}
\alias{launch_shinytmb}
\title{A high level wrapper to launch shinystan for a TMB fit.}
\usage{
launch_shinytmb(fit)
}
\arguments{
\item{fit}{A named list returned by \code{sample_tmb}.}
}
\description{
A high level wrapper to launch shinystan for a TMB fit.
}
\details{
This function simply calls
  \code{launch_shinystan(as.shinystan.tmb(tmb.fit))}.
}
