% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_admb.R
\name{pairs_admb}
\alias{pairs_admb}
\title{Deprecated function to make custom pairs plots for 'adfit'
objects. Use S3 class method 'pairs' instead, and see
\code{?pairs.adfit} for help.}
\usage{
pairs_admb(...)
}
\arguments{
\item{...}{Passed on}
}
\description{
Deprecated function to make custom pairs plots for 'adfit'
objects. Use S3 class method 'pairs' instead, and see
\code{?pairs.adfit} for help.
}
