% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_marginals}
\alias{plot_marginals}
\title{Plot marginal distributions for a fitted model}
\usage{
plot_marginals(fit, mon = NULL)
}
\arguments{
\item{fit}{A fitted object returned by
\code{\link{sample_admb}}.}

\item{mon}{An object returned by rstan::monitor(fit)}
}
\description{
Plot marginal distributions for a fitted model
}
\details{
This function plots 4x4 grid cells of all parameters
  in a model, comparing the marginal posterior histogram vs
  the asympotitic normal (red lines) from the inverse
  Hessian. Its intended use is to quickly gauge differences
  between frequentist and Bayesian inference on the same
  model.

If argument \code{mon} is provided, the effective sample size
(ESS) and R-hat estimates are printed in the top right
corner. See
\url{https://mc-stan.org/rstan/reference/Rhat.html} for more
information. Generally Rhat>1.05 or ESS<100 (per chain)
suggest inference may be unreliable.

This function is customized to work with multipage PDFs,
specifically:
\code{pdf('marginals.pdf', onefile=TRUE, width=7,height=5)}
}
\examples{
fit <- readRDS(system.file('examples', 'fit_admb.RDS', package='adnuts'))
mon <- rstan::monitor(fit$samples, warmup=fit$warmup, print=FALSE)
plot_marginals(fit, mon)

}
