% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_sampler_params}
\alias{plot_sampler_params}
\title{Plot adaptation metrics for a fitted model.}
\usage{
plot_sampler_params(fit, plot = TRUE)
}
\arguments{
\item{fit}{A fitted object returned by
\code{\link{sample_admb}}.}

\item{plot}{Whether to plot the results}
}
\value{
Prints and invisibly returns a ggplot object
}
\description{
Plot adaptation metrics for a fitted model.
}
\details{
This utility function quickly plots the adaptation output of NUTS
chains.
}
\examples{
fit <- readRDS(system.file('examples', 'fit_admb.RDS',
  package='adnuts'))
plot_sampler_params(fit)
}
