% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.adfit}
\alias{print.adfit}
\title{Print summary of adfit object}
\usage{
\method{print}{adfit}(x, ...)
}
\arguments{
\item{x}{Fitted object from \code{\link{sample_admb}}}

\item{...}{Ignored}
}
\value{
Summary printed to console
}
\description{
Print summary of adfit object
}
