% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_mle_fit}
\alias{read_mle_fit}
\title{Read maximum likelihood fit for ADMB model}
\usage{
read_mle_fit(model, path = getwd())
}
\arguments{
\item{model}{Model name}
}
\value{
A list containing, MLE estimates, standard errors, covariance
  and correlation matrices, and other output from ADMB.
}
\description{
Read maximum likelihood fit for ADMB model
}
\details{
This is based loosely off read.admbFit from r4ss.
}

