% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_psv}
\alias{read_psv}
\title{Read in an ADMB PSV file}
\usage{
read_psv(fn, names = NULL)
}
\arguments{
\item{fn}{The filename (with or without the .psv extension)}

\item{names}{Column names for the output. If `NULL`, they will be sequential
starting with the letter 'V'}
}
\value{
A data frame containing the PSV file output
}
\description{
Read in an ADMB PSV file
}
