% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_admb.R
\name{sample_admb}
\alias{sample_admb}
\title{Sample from and ADMB object, using the NUTS or RWM algorithms.}
\usage{
sample_admb(model, iter, init, chains = 1, warmup = NULL, seeds = NULL,
  thin = 1, dir = getwd(), mceval = FALSE, duration = NULL,
  parallel = FALSE, cores = NULL, control = NULL, algorithm = "NUTS",
  ...)
}
\arguments{
\item{model}{Name of model}

\item{iter}{Total iterations to run.}

\item{init}{Initial values. Can be NULL (use MLE), or a list of vectors,
one for each chain.}

\item{chains}{The number of chains to run.}

\item{warmup}{The number of warmup samples.}

\item{seeds}{Random number seeds one for each chain.}

\item{thin}{The thinning rate.}

\item{dir}{The name of a folder containing the ADMB model, which should
not not be the working directory. This function requires this for
parallel since the folder is copied and run in parallel.}

\item{mceval}{Whether to run the model with \code{-mceval} afterward.}

\item{duration}{The number of minutes after which the model will quit
running.}

\item{parallel}{Whether to run chains in parallel.}

\item{cores}{If parallel is \code{TRUE}, how many cores to use.}

\item{control}{A list of control options for the algorithms. See
\code{sample_tmb} for more information.}

\item{algorithm}{Which algorithm to use, either "NUTS" or "RWM".}
}
\description{
Sample from and ADMB object, using the NUTS or RWM algorithms.
}

