% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_admb.R
\name{sample_admb}
\alias{sample_admb}
\title{Deprecated version of wrapper function. Use sample_nuts or
sample_rwm instead.}
\usage{
sample_admb(
  model,
  path = getwd(),
  iter = 2000,
  init = NULL,
  chains = 3,
  warmup = NULL,
  seeds = NULL,
  thin = 1,
  mceval = FALSE,
  duration = NULL,
  parallel = FALSE,
  cores = NULL,
  control = NULL,
  skip_optimization = TRUE,
  algorithm = "NUTS",
  skip_monitor = FALSE,
  skip_unbounded = TRUE,
  admb_args = NULL,
  ...
)
}
\arguments{
\item{model}{Name of model (i.e., model.tpl)}

\item{path}{Path to model executable. Defaults to working
directory. Often best to have model files in a separate
subdirectory, particularly for parallel.}

\item{iter}{The number of samples to draw.}

\item{init}{A list of lists containing the initial parameter
vectors, one for each chain or a function. It is strongly
recommended to initialize multiple chains from dispersed
points. A of NULL signifies to use the starting values
present in the model (i.e., \code{obj$par}) for all chains.}

\item{chains}{The number of chains to run.}

\item{warmup}{The number of warmup iterations.}

\item{seeds}{A vector of seeds, one for each chain.}

\item{thin}{The thinning rate to apply to samples. Typically
not used with NUTS.}

\item{mceval}{Whether to run the model with \code{-mceval} on
samples from merged chains.}

\item{duration}{The number of minutes after which the model
will quit running.}

\item{parallel}{A deprecated argument, use cores=1 for serial
execution or cores>1 for parallel (default is to parallel
with cores equal to the available-1)}

\item{cores}{The number of cores to use for parallel
execution.}

\item{control}{A list to control the sampler. See details for
further use.}

\item{skip_optimization}{Whether to run the optimizer before
running MCMC. This is rarely need as it is better to run it
once before to get the covariance matrix, or the estimates
are not needed with adaptive NUTS.}

\item{algorithm}{The algorithm to use."NUTS" is the default
and recommended one, but "RWM" for the random walk
Metropolis sampler and "HMC" for the static HMC sampler are
available. HMC is deprecated but may be of use in special
situations. These algorithms require different arguments;
see their help files for more information.}

\item{skip_monitor}{Whether to skip calculating diagnostics
(effective sample size, Rhat) via the \code{rstan::monitor}
function. This can be slow for models with high dimension or
many iterations. The result is used in plots and summaries
so it is recommended to turn on. If model run with
\code{skip_monitor=FALSE} you can recreate it post-hoc by
setting \code{fit$monitor=rstan::monitor(fit$samples,
fit$warmup, print=FALSE)}.}

\item{skip_unbounded}{Whether to skip returning the unbounded
version of the posterior samples in addition to the bounded
ones. It may be advisable to set to FALSE for very large
models to save space.}

\item{admb_args}{A character string which gets passed to the
command line, allowing finer control}

\item{...}{Further arguments to be passed to the
algorithm. See help files for the samplers for further
arguments.}
}
\description{
Deprecated version of wrapper function. Use sample_nuts or
sample_rwm instead.
}
\section{Warning}{
 This is deprecated and will cease to exist
  in future releases
}

