% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{sample_admb_parallel}
\alias{sample_admb_parallel}
\title{A wrapper for running SS models in parallel}
\usage{
sample_admb_parallel(parallel_number, dir, ...)
}
\description{
A wrapper for running SS models in parallel
}

