% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_admb.R
\name{sample_admb_rwm}
\alias{sample_admb_rwm}
\title{Run a single random walk Metropolis chain for an ADMB model}
\usage{
sample_admb_rwm(
  path,
  model,
  iter = 2000,
  thin = 1,
  warmup = ceiling(iter/2),
  init = NULL,
  chain = 1,
  seed = NULL,
  control = NULL,
  verbose = TRUE,
  duration = NULL,
  extra.args = NULL,
  admb_args = NULL,
  skip_optimization = TRUE
)
}
\arguments{
\item{path}{Path to model executable. Defaults to working
directory. Often best to have model files in a separate
subdirectory, particularly for parallel.}

\item{model}{Name of model (i.e., model.tpl)}

\item{iter}{The number of samples to draw.}

\item{thin}{The thinning rate to apply to samples. Typically
not used with NUTS.}

\item{warmup}{The number of warmup iterations.}

\item{init}{A list of lists containing the initial parameter
vectors, one for each chain or a function. It is strongly
recommended to initialize multiple chains from dispersed
points. A of NULL signifies to use the starting values
present in the model (i.e., \code{obj$par}) for all chains.}

\item{control}{A list to control the sampler. See details for
further use.}

\item{duration}{The number of minutes after which the model
will quit running.}

\item{admb_args}{A character string which gets passed to the
command line, allowing finer control}

\item{skip_optimization}{Whether to run the optimizer before
running MCMC. This is rarely need as it is better to run it
once before to get the covariance matrix, or the estimates
are not needed with adaptive NUTS.}
}
\description{
A low level function to run a single chain. Unlikely to be used by a
user, instead prefer \code{\link{sample_admb}}
}
\seealso{
\code{\link{sample_admb}}
}
