% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{sample_sparse_tmb}
\alias{sample_sparse_tmb}
\title{Fit a TMB model using a sparse inverse mass matrix}
\usage{
sample_sparse_tmb(
  obj,
  iter,
  warmup,
  cores,
  chains,
  control = NULL,
  seed = NULL,
  init = c("last.par.best", "random"),
  metric = c("sparse", "dense", "diag", "unit"),
  skip_optimization = FALSE,
  Q = NULL,
  Qinv = NULL,
  globals = NULL,
  ...
)
}
\arguments{
\item{obj}{The TMB object with random effects turned on and
optimized}

\item{iter}{Total iterations to run (warmup + sampling)}

\item{warmup}{Total warmup iterations}

\item{cores}{Number of parallel cores to use}

\item{chains}{Number of chains}

\item{control}{NUTS control list, currently available options
are 'adapt_delta', 'max_treedepth', and 'metric' which is
the type of metric adaptation for Stan to do with options
('unit_e', 'diag_e', or 'dense_e'). For dense and sparse
metrics this usually can be 'unit_e' to skip adaptation.
NULL values (default) revert to \code{stan_sample} defaults.}

\item{seed}{Random number seed}

\item{init}{Either 'last.par.best' (default) or 'random'. The
former starts from the joint mode and the latter draws from
a multivariate t distribution with df=1 centered at the mode
using the inverse joint precision matrix as a covariance
matrix. Note that StanEstimators only allows for the same
init vector for all chains currently. If a seed is specified
it will be set and thus the inits used will be reproducible.}

\item{skip_optimization}{Whether to skip optimization or not
(default).}

\item{globals}{A named list of objects to pass to new R
sessions when running in parallel and using RTMB. Typically
this is the `data` object for now.}

\item{...}{Additional arguments to pass to
\code{\link{StanEstimators::stan_sample}}.}
}
\value{
A fitted MCMC object of class 'adfit'
}
\description{
Fit a TMB model using a sparse inverse mass matrix
}
