% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{sample_sparse_tmb}
\alias{sample_sparse_tmb}
\title{Fit a TMB model using a sparse inverse mass matrix}
\usage{
sample_sparse_tmb(
  obj,
  iter,
  warmup,
  cores,
  chains,
  control = NULL,
  seed = NULL
)
}
\arguments{
\item{obj}{The TMB object with random effects turned on and
optimized}

\item{iter}{Total iterations to run (warmup + sampling)}

\item{warmup}{Total warmup iterations}

\item{cores}{Number of parallel cores to use}

\item{chains}{Number of chains}

\item{control}{NUTS control list}

\item{seed}{Random number seed}
}
\value{
A fitted MCMC object of class 'adfit'
}
\description{
Fit a TMB model using a sparse inverse mass matrix
}
