% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_tmb.R
\name{sample_tmb}
\alias{sample_tmb}
\title{MCMC sampling of TMB models}
\usage{
sample_tmb(obj, iter, init, chains = 1, seeds = NULL, lower = NULL,
  thin = 1, upper = NULL, control = NULL, ...)
}
\arguments{
\item{obj}{A TMB model object.}

\item{iter}{The number of samples to draw.}

\item{init}{A list of lists containing the initial parameter vectors,
one for each chain or a function. It is strongly recommended to
initialize multiple chains from dispersed points. A of NULL signifies
to use the starting values present in the model (i.e., \code{obj$par})
for all chains.}

\item{control}{A list to control the sampler. Elements are
\itemize{
\item{"adapt_delta"}{The target acceptance rate.}
\item{"metric"}{ The mass metric to use. Options are:}
\itemize{
\item{"unit"}{ for a unit diagonal matrix}
\item{"diag"}{to estimate a diagonal matrix during warmup}
\item{"matrix"} a matrix to be used directly (in untransformed space). For
  instance from the MLE Hessian.
}
\item{"algorithm"}{ A string specifiying an algorithm. Currently supported
  are: \itemize{ \item{"RWM"}{the random walk Metropolis sampler}
  \item{"HMC"}{the static HMC sampler}
  \item{"NUTS"}{the No-U-Turn sampler} }
  These algorithms require different arguments; see their help files for
  more information.}
}
\item{"adapt_engaged"}{Whether adaptation of step size and metric is turned on}
\item{"thin"}{Thin The thinning rate to apply to samples. The default of 1
 is almost always best.}}

\item{...}{Further arguments to be passed to the algorithm. See help
files for the samplers for further arguments.}
}
\value{
A list containing the samples,  and properties of the sampler
  useful for diagnosing behavior and efficiency.
}
\description{
[BETA VERSION] Draw samples from the posterior of a TMB model using a
specified MCMC algorithm.
}
\details{
This function is a top-level wrapper designed specifically to
  work with TMB models. There are several MCMC algorithms available for
  use. The user is responsible for specifying the model properly
  (priors, starting values, desired parameters fixed, etc.), as well as
  assessing the convergence and validity of the resulting samples (e.g.,
  through the \code{coda} package) before making inference.
}
\author{
Cole Monnahan
}
\seealso{
\code{\link{extract_samples}}, \code{\link{launch_shinystan_tmb}}
}

