% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_tmb.R
\name{sample_tmb}
\alias{sample_tmb}
\title{MCMC sampling of TMB models}
\usage{
sample_tmb(obj, iter, init, chains = 1, seeds = NULL, lower = NULL,
  upper = NULL, control = NULL, ...)
}
\arguments{
\item{obj}{A TMB model object.}

\item{iter}{The number of samples to draw.}

\item{init}{A list of lists containing the initial parameter vectors,
one for each chain or a function. It is strongly recommended to
initialize multiple chains from dispersed points. A of NULL signifies
to use the starting values present in the model (i.e., \code{obj$par})
for all chains.}

\item{...}{Further arguments to be passed to the algorithm. See help
files for the samplers for further arguments.}
}
\value{
A list containing the samples,  and properties of the sampler
  useful for diagnosing behavior and efficiency.
}
\description{
[BETA VERSION] Draw samples from the posterior of a TMB model using a
specified MCMC algorithm.
}
\details{
This function is a top-level wrapper designed specifically to
  work with TMB models. There are several MCMC algorithms available for
  use. The user is responsible for specifying the model properly
  (priors, starting values, desired parameters fixed, etc.), as well as
  assessing the convergence and validity of the resulting samples (e.g.,
  through the \code{coda} package) before making inference.
}
\author{
Cole Monnahan
}
\seealso{
\code{\link{extract_samples}}, \code{\link{launch_shinystan_tmb}}
}

