% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-admb-cov.R
\name{write_admb_cov}
\alias{write_admb_cov}
\title{Write a covariance matrix to \code{admodel.cov}.}
\usage{
write_admb_cov(path = NULL, cov_unbounded, hbf = NULL, fn_cov = "admodel.cov")
}
\arguments{
\item{path}{Path to model.}

\item{cov_unbounded}{The covariance matrix in unbounded space.}

\item{hbf}{The hybrid_bounded_flag value. Use \code{hbf = 1} for HMC.}

\item{fn_cov}{Name of the ADMB covariance file}
}
\description{
Write a covariance matrix to \code{admodel.cov}.
}
