% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/write-psv.R
\name{write_psv}
\alias{write_psv}
\title{Write matrix of samples to a binary .psv file.}
\usage{
write_psv(path, fn_psv, samples)

write_psv(path, fn_psv, samples)
}
\arguments{
\item{path}{Directory to write the PSV file in}

\item{fn_psv}{Model PSV file name with or without the PSV extension}

\item{samples}{A matrix or data.frame of samples, each column is a
parameter, each row a sample.}

\item{fn}{Model executable name}
}
\description{
Write matrix of samples to a binary .psv file.

Write an ADMB PSV file
}
\details{
Useful to combine multiple MCMC runs together into a single
.psv file which can then be executed with \code{-mceval}.

Useful to combine multiple MCMC runs together into a single
\verb{.psv file} which can then be executed with '-mceval'.
}
